feat: ref. big query module
