+++ src/big-query-support/pull-data/getLastProcessedBlock.ts
@@ -1,10 +1,9 @@
-/* eslint-disable @typescript-eslint/no-unsafe-argument */
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 
 import { BigQuery } from '@google-cloud/bigquery';
 
-import { LAST_PROCESSED_BLOCK_TABLE_ID } from '../common/constants';
+import { LAST_PROCESSED_BLOCK_TABLE_ID } from '../../common';
+
 
 export const getLastProcessedBlock = async (
   bigQuery: BigQuery,
