--- src/big-query-support/pullExistingPositionRow.ts
@@ -1,96 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unsafe-argument */
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-
-import { BigQuery } from '@google-cloud/bigquery';
-
-import { POSITIONS_TABLE_ID } from '../common';
-import { bqNumericToNumber, bqTimestampToUnixSeconds } from './utils';
-
-export type BigQueryPositionRow = {
-  marginEngineAddress: string; // immutable
-  vammAddress: string; // immutable
-  ownerAddress: string; // immutable
-  tickLower: number; // immutable
-  tickUpper: number; // immutable
-  realizedPnLFromSwaps: number;
-  realizedPnLFromFeesPaid: number;
-  netNotionalLocked: number;
-  netFixedRateLocked: number;
-  lastUpdatedBlockNumber: number;
-  notionalLiquidityProvided: number;
-  realizedPnLFromFeesCollected: number;
-  netMarginDeposited: number;
-  rateOracleIndex: number; // immutable
-  rowLastUpdatedTimestamp: number;
-  fixedTokenBalance: number;
-  variableTokenBalance: number;
-  positionInitializationBlockNumber: number; // immutable
-  rateOracle: string; // immutable
-  underlyingToken: string; // immutable
-  chainId: number; // immutable
-  cashflowLiFactor: number;
-  cashflowTimeFactor: number;
-  cashflowFreeTerm: number;
-  tickPrevious: number;
-  liquidity: number;
-};
-
-export const pullExistingPositionRow = async (
-  bigQuery: BigQuery,
-  chainId: number,
-  vammAddress: string,
-  recipient: string,
-  tickLower: number,
-  tickUpper: number,
-): Promise<BigQueryPositionRow | null> => {
-  const sqlQuery = `
-    SELECT * FROM \`${POSITIONS_TABLE_ID}\` 
-      WHERE chainId=${chainId} AND
-            vammAddress=\"${vammAddress}\" AND 
-            ownerAddress=\"${recipient}\" AND 
-            tickLower=${tickLower} AND 
-            tickUpper=${tickUpper}
-  `;
-
-  const options = {
-    query: sqlQuery,
-  };
-
-  const [rows] = await bigQuery.query(options);
-
-  if (!rows || rows.length === 0) {
-    return null;
-  }
-
-  // todo: operations like this one need validation
-
-  return {
-    marginEngineAddress: rows[0].marginEngineAddress,
-    vammAddress: rows[0].vammAddress,
-    ownerAddress: rows[0].ownerAddress,
-    tickLower: rows[0].tickLower,
-    tickUpper: rows[0].tickUpper,
-    realizedPnLFromSwaps: bqNumericToNumber(rows[0].realizedPnLFromSwaps),
-    realizedPnLFromFeesPaid: bqNumericToNumber(rows[0].realizedPnLFromFeesPaid),
-    netNotionalLocked: bqNumericToNumber(rows[0].netNotionalLocked),
-    netFixedRateLocked: bqNumericToNumber(rows[0].netFixedRateLocked),
-    lastUpdatedBlockNumber: rows[0].lastUpdatedBlockNumber,
-    notionalLiquidityProvided: bqNumericToNumber(rows[0].notionalLiquidityProvided),
-    realizedPnLFromFeesCollected: bqNumericToNumber(rows[0].realizedPnLFromFeesCollected),
-    netMarginDeposited: bqNumericToNumber(rows[0].netMarginDeposited),
-    rateOracleIndex: bqNumericToNumber(rows[0].rateOracleIndex),
-    rowLastUpdatedTimestamp: bqTimestampToUnixSeconds(rows[0].rowLastUpdatedTimestamp),
-    fixedTokenBalance: bqNumericToNumber(rows[0].fixedTokenBalance),
-    variableTokenBalance: bqNumericToNumber(rows[0].variableTokenBalance),
-    positionInitializationBlockNumber: rows[0].positionInitializationBlockNumber,
-    rateOracle: rows[0].rateOracle,
-    underlyingToken: rows[0].underlyingToken,
-    chainId: rows[0].chainId,
-    cashflowLiFactor: bqNumericToNumber(rows[0].cashflowLiFactor),
-    cashflowTimeFactor: bqNumericToNumber(rows[0].cashflowTimeFactor),
-    cashflowFreeTerm: bqNumericToNumber(rows[0].cashflowFreeTerm),
-    liquidity: bqNumericToNumber(rows[0].liquidity),
-    tickPrevious: rows[0].tickPrevious,
-  };
-};
