+++ src/big-query-support/utils.ts
@@ -1,9 +1,6 @@
 import { BigQuery, BigQueryInt, BigQueryTimestamp } from '@google-cloud/bigquery';
 
 export const bqNumericToNumber = (bqNumeric: BigQueryInt): number => {
-  if (!bqNumeric) {
-    console.log('Bq numberic undefined!!!!');
-  }
   return Number(bqNumeric.toString());
 };
