+++ src/common/types.ts
@@ -1,8 +1,10 @@
 import { AMM } from '@voltz-protocol/v1-sdk';
 import { Event } from 'ethers';
 
+export type EventType = 'mint' | 'burn' | 'swap' | 'price_change';
+
 export interface ExtendedEvent extends Event {
-  type: 'mint' | 'burn' | 'swap' | 'price_change';
   amm: AMM;
   chainId: number;
-}
+  type: EventType;
+};
